Title Mplus Boostrapping

Title Mplus Boostrapping

MplusBootstrap(
  Model,
  type,
  first_path,
  second_path = NULL,
  multi_level,
  iv = NULL,
  m = NULL,
  w = NULL,
  xw = NULL,
  dv = NULL,
  first_stage_mod = NULL,
  level_one_mod = NULL,
  output = "Untitled",
  rep = 20000,
  conf = 95
)

Arguments

Model

a list object containing the parameters of the model (from Mplus output file)

type

should choose from "med" "mod-med" or "med-mod"

first_path

the name of the first random path, left as 'fixed' when the path is not random (must be specified by the user)

second_path

the name of the second random path, left as 'fixed' when the path is not random (must be specified by the user)

multi_level

a binary variable indicating whether the specified is multi-level (must be specified by the user)

iv

the name of the independent variable

m

the name of the mediator

w

the name of the moderator

xw

the name of the interaction term

dv

the name of the dependent variable

first_stage_mod

a binary variable indicating whether the moderation effect is on the first path

level_one_mod

a binary variable indicating whether the moderation effect is on the first level

output

the name of the output image file (.png)

rep

the number of repetitions (default: 20000)

conf

the critical value used to generate confidence interval

Value

An image containing the distribution of the indirect effect and confidence interval