Title: Moderated Mediation test of multilevel model (random-random; moderator at level two) This function can calculate moderated mediation effect of multi-level model that contains two random paths.

Title: Moderated Mediation test of multilevel model (random-random; moderator at level two) This function can calculate moderated mediation effect of multi-level model that contains two random paths.

complex_modmed(
  Model,
  first_stage_mod = TRUE,
  first_path = NULL,
  second_path = NULL,
  w = NULL,
  output = "Untitled",
  rep = 20000,
  conf = 95
)

Arguments

Model

list object containing the parameters of the model (from Mplus output file)

first_stage_mod

a binary variable indicating whether the moderation effect is on the first path (default: TRUE)

first_path

the name of the first random path (should not be 'fixed')

second_path

the name of the second random path (should not be 'fixed')

w

the name of the moderator

output

the name of the output image file (.png)

rep

the number of repetitions (default: 20000)

conf

the critical value used to generate confidence interval

Value

An image containing the distribution of the indirect effect and confidence interval