Title: Complex version of mediation test

The template of the monte carlo method in this function is from https://quantpsy.org/medmc/medmc.htm. This function can handle multi-level model (1-1-1 mediation with two random paths).

complex_version(
  Model,
  first_path = NULL,
  second_path = NULL,
  output = "Untitled",
  rep = 20000,
  conf = 95
)

Arguments

Model

a list object containing the parameters of the model (from Mplus output file)

first_path

the name of the first random path (should not be 'fixed')

second_path

the name of the second random path (should not be 'fixed')

output

the name of the output image file (.png)

rep

the number of repetitions (default: 20000)

conf

the critical value used to generate confidence interval

Value

An image containing the distribution of the indirect effect and confidence interval

Examples

# 1-1-1 mediation with two random paths
model <- load_model('inst/examples/random_random')
#> Error in convert_to_filelist(target, filefilter, recursive): Cannot find target: inst/examples/random_random.inp
simple_version(model, 's1', 's2', output='random_random')
#> Error: object 'model' not found