Title: Simple version of mediation test for mediated moderation test

The template of the monte carlo method in this function is from https://quantpsy.org/medmc/medmc.htm. This function can handle multi-level mediated moderation model where the moderator can be specified at either level one or two.

med_mod(
  Model,
  iv = NULL,
  m = NULL,
  w = NULL,
  xw = NULL,
  dv = NULL,
  first_path = NULL,
  level_one_mod = TRUE,
  output = "Untitled",
  rep = 20000,
  conf = 95
)

Arguments

Model

a list object containing the parameters of the model (from Mplus output file)

iv

the name of the independent variable

m

the name of the mediator

w

the name of the moderator

xw

the name of the interaction term

dv

the name of the dependent variable

first_path

the name of the first random path (should not be 'fixed')

level_one_mod

a binary variable indicating whether the moderation effect is on the first level (default: TRUE)

output

the name of the output image file (.png)

rep

the number of repetitions (default: 20000)

conf

the critical value used to generate confidence interval

Value

An image containing the distribution of the indirect effect and confidence interval